17-06

+++ webcoso/Controllers/NhomSanPhamsController.cs
@@ -8,16 +8,20 @@ using System.Web;
 using System.Web.Mvc;
 using webcoso.Models;
 using PagedList;
+using webcoso.Message;
 
 namespace webcoso.Controllers
 {
     public class NhomSanPhamsController : Controller
     {
         private WebcosoContext db = new WebcosoContext();
+        private ApplicationDbContext data = new ApplicationDbContext();
 
         // GET: NhomSanPham
         public ActionResult Index(int? page)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             if (page == null) page = 1;
             var all_sach = (from s in db.NhomSanPham select s).OrderBy(m => m.MaNhom);
             int pageSize = 5;
@@ -28,6 +32,8 @@ namespace webcoso.Controllers
         // GET: NhomSanPham/Details/5
         public ActionResult Details(int? id)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -43,6 +49,8 @@ namespace webcoso.Controllers
         // GET: NhomSanPham/Create
         public ActionResult Create()
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             return View();
         }
 
@@ -53,6 +61,8 @@ namespace webcoso.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "MaNhom,TenNhom")] NhomSanPham nhomSanPham)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             if (ModelState.IsValid)
             {
                 db.NhomSanPham.Add(nhomSanPham);
@@ -66,6 +76,8 @@ namespace webcoso.Controllers
         // GET: NhomSanPham/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -85,6 +97,8 @@ namespace webcoso.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "MaNhom,TenNhom")] NhomSanPham nhomSanPham)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             if (ModelState.IsValid)
             {
                 db.Entry(nhomSanPham).State = EntityState.Modified;
@@ -97,6 +111,8 @@ namespace webcoso.Controllers
         // GET: NhomSanPham/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -114,7 +130,15 @@ namespace webcoso.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             NhomSanPham nhomSanPham = db.NhomSanPham.Find(id);
+            if (db.SanPham.Where(p => p.MaNhom == nhomSanPham.MaNhom).FirstOrDefault() != null)
+            {
+                Notification.set_flash("Không thể xoá nhóm \' " + nhomSanPham.TenNhom + " \'!", "error");
+                return RedirectToAction("Index");
+            }
+            Notification.set_flash("Đã xoá nhóm \' " + nhomSanPham.TenNhom + " \'!", "error");
             db.NhomSanPham.Remove(nhomSanPham);
             db.SaveChanges();
             return RedirectToAction("Index");
@@ -127,6 +151,19 @@ namespace webcoso.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+
+        public bool AuthAdmin()
+        {
+            var user = data.Users.FirstOrDefault(u => u.UserName == User.Identity.Name);
+            if (user == null)
+                return false;
+            var userExist = user.Roles.FirstOrDefault(r => r.UserId == user.Id);
+            if (userExist == null)
+                return false;
+            if (userExist.RoleId != "1")
+                return false;
+            return true;
         }
     }
 }
